requirements.txt
pillow
numpy
selenium
mss
